(PC-19501) refactor(bookOffer): extract BookingConfirmation and BookingOfferModal tests from __tests__ folder
